% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sicure.R
\name{fun.opt}
\alias{fun.opt}
\title{Objective function}
\usage{
fun.opt(theta_h, x_cov, time, delta)
}
\arguments{
\item{theta_h}{A numeric vector containing the initial iterant for the vector of parameters and the initial bandwidths.}

\item{x_cov}{A matrix or data frame giving the covariate values. Each row represents an individual and each column corresponds to a variable.}

\item{time}{A numeric vector giving the observed times.}

\item{delta}{A numeric vector giving the values of the uncensoring indicator, where 1 indicates that the event of interest has been observed and 0 indicates that the observation is censored.}
}
\value{
The value of the negative log-likelihood.
}
\description{
This function computes the negative log-likelihood for a given set of parameters.
}
